+++ CHANGELOG.md
@@ -5,7 +5,6 @@
 * scores.clj
 * stocks.clj
 * admin/toggle-status
-* time restrictions
 * digest namespace. used only from answers
 * effective upsert
 * emulate flash
@@ -15,10 +14,19 @@
 
 # 0.3.0-SNAPSHOT
 
-* restrictions
+- java25 on macos
+- restrictions
+    - system/min-interval-answers
+    - system/min-interval-comments
+    - system/min-interval-uploads
+    - system/max-comments
+    - system/max-uploads
+    - system/kp2-flash
+- restrict comments
+* restrict uploads
 * improve admin page
 * padding y in tasks page
-- java25 on macos
+
 
 # 0.2.16 (2025-09-17)